fix: rename useRegister on useAuth, add login function in useAuth
